requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
tensorflow==2.8.0
keras==2.8.0
numpy==1.21.0
opencv-python-headless==4.5.5.64
face-recognition==1.3.0
Pillow==8.4.0
gunicorn==20.1.0
torch==1.9.0
torchvision==0.10.0